Refactor build process to use a faster, more efficient method
The old build method required a mess of shell scripts and duct
tape. The new method uses Docker build args to parameterize
creating a build container and creates a zip archive of the
collected dependencies. The zip can be retained to cache builds
and is named the shasum of the requirements file in order to
prevent unnecessary building. The zip is extracted to the
`install_dir` to prepare for packaging (either manually, or via
lambda-uploader.

Because of the scope of things, there are some breaking changes:
* No more wheelhouse because we don't have to build wheels.
* wheelhouse is now dependency_cache in yoke.yml.
* Default dependency_cache is now ~/.yoke/dependencies
  because this makes more sense than an arbitrary parent path.
* install_dir is now defaulted to 'dist' instead of 'lib' to
  be more specific of the purpose of the directory.
* 'openssl', 'libffi', and 'libxml' are no longer dependencies
  options since they can be installed via packages now.