fix(templates): pin auditwheel to 1.9
A recent release (November 2018) of auditwheel was accepted in December
2018 into the manylinux container. For whatever reason, that new version
of auditwheel causes pyyaml to never get output from the build
container.

This took a while to find due to the semantics and the silent `rm` of
the bash one-liner that we're using to finalize the wheels. It's not
clear to me why we're removing wheel files (see
https://github.com/rackerlabs/yoke/blob/00c01d6217b77f9ce14233a0f3686ac6f7b1c247/yoke/templates.py#L229-L233)
and I need to dig into this further. However, for the moment, I just
need this change to get some builds unstuck. I've confirmed pinning
auditwheel does fix the issue.